backend_federal_plan/requirements.txt
fastapi==0.128.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
uvicorn==0.40.0
python-dotenv~=1.0.1
aiogram~=3.18.0
SQLAlchemy~=2.0.38
aiosqlite~=0.21.0
pytest~=8.3.5
pytest-asyncio~=0.25.3